TT-1990 Use Correct Role
* When looking up role, filter for whether we are online.
* Refactor all role logic to custom hook
