+++ src/components/GroupSettings/useReviewerIds.tsx
@@ -1,19 +1,19 @@
 import { useGlobal } from 'reactn';
-import { GroupMembership, Role, RoleNames } from '../../model';
-import { related, getRoleId } from '../../crud';
+import { GroupMembership, RoleNames } from '../../model';
+import { related, useRole } from '../../crud';
 import { IPerson } from './TeamCol';
 
 interface IProps {
   groupMemberships: GroupMembership[];
-  roles: Role[];
 }
 
 function useReviewerIds(props: IProps): IPerson[] {
-  const { groupMemberships, roles } = props;
+  const { groupMemberships } = props;
+  const { getRoleId } = useRole();
   const [group] = useGlobal('group');
 
-  const transcriberId = getRoleId(roles, RoleNames.Transcriber);
-  const editorId = getRoleId(roles, RoleNames.Editor);
+  const transcriberId = getRoleId(RoleNames.Transcriber);
+  const editorId = getRoleId(RoleNames.Editor);
 
   return groupMemberships
     .filter(
