+++ src/components/GroupSettings/useTranscriberIds.tsx
@@ -1,18 +1,18 @@
 import { useGlobal } from 'reactn';
-import { GroupMembership, Role, RoleNames } from '../../model';
-import { related, getRoleId } from '../../crud';
+import { GroupMembership, RoleNames } from '../../model';
+import { related, useRole } from '../../crud';
 import { IPerson } from './TeamCol';
 
 interface IProps {
   groupMemberships: GroupMembership[];
-  roles: Role[];
 }
 
 function useTranscriberIds(props: IProps): IPerson[] {
-  const { groupMemberships, roles } = props;
+  const { groupMemberships } = props;
+  const { getRoleId } = useRole();
   const [group] = useGlobal('group');
 
-  const transcriberId = getRoleId(roles, RoleNames.Transcriber);
+  const transcriberId = getRoleId(RoleNames.Transcriber);
 
   return groupMemberships
     .filter((gm) => related(gm, 'group') === group)
